third_party/blink/web_tests/http/tests/devtools/elements/styles-2/keyframes-rules-expected.txt

Tests that source data is extracted correctly from stylesheets with @keyframes rules.


Running: testInit

Running: testDumpStyles
[expanded] 
element.style { ()
    background-color: white;

[expanded] 
* { (<style>)
/-- overloaded --/     background-color: papayawhip;

[expanded] 
div { (user agent stylesheet)
    display: block;
    unicode-bidi: isolate;
